Item structure and adding function.
Structure of the single tree element was created.
Function that adds new item to the tree was created and it works properly.

Simple function that displays every node of the tree in order of their creation was writen and works but needs to be changed because of that creation order doesn't show the relations between nodes of te tree.

To do list updated.